float32 -> float64

+++ dortislem.go
@@ -12,6 +12,6 @@ func Cikar(x int, y int) int {
 	return x - y
 }
 
-func Bol(x int, y int) float32 {
+func Bol(x int, y int) float64 {
 	return x / y
 }
